Lifted filter state up, so that it persists when switching sidebar tabs
