+++ src/components/AnnotationDesktop/LayersPanel/FilterSettings/index.ts
@@ -1,1 +1,2 @@
 export * from './FilterSettings';
+export * from './FilterState';
