+++ src/components/AnnotationDesktop/LayersPanel/LayersPanel.tsx
@@ -32,7 +32,6 @@ export const LayersPanel = (props: LayersPanelProps) => {
       <FilterSettings 
         i18n={props.i18n}
         layers={props.layers}
-        present={props.present} 
         onChangeFilter={props.onChangeFilter} />
     </div>
   )
