+++ src/components/AnnotationDesktop/ViewMenu/ViewMenu.tsx
@@ -12,6 +12,7 @@ import { LayersPanel } from '../LayersPanel';
 import { DocumentNotes, DocumentNotesList, DocumentNotesMenuIcon } from '../DocumentNotes';
 
 import './ViewMenu.css';
+import { FilterState } from '../LayersPanel/FilterSettings';
 
 interface ViewMenuProps {
 
@@ -83,80 +84,82 @@ export const ViewMenu = (props: ViewMenuProps) => {
       className="anno-sidebar-container"
       data-collapsed={panel ? undefined : 'true'}>
 
-      <DocumentNotes
-        channelId={props.channel}
-        layerId={props.defaultLayer}
-        present={props.present}
-        onError={onError}>
+      <FilterState present={props.present}>
+        <DocumentNotes
+          channelId={props.channel}
+          layerId={props.defaultLayer}
+          present={props.present}
+          onError={onError}>
 
-        <div 
-          className="anno-menubar anno-desktop-overlay view-menu">
-          <section>
-            <button 
-              className={panel === ViewMenuPanel.ANNOTATIONS ? 'active' : undefined}
-              onClick={() => togglePanel(ViewMenuPanel.ANNOTATIONS)}>
-              <Chats />
-            </button>
-
-            <button
-              className={panel === ViewMenuPanel.LAYERS ? 'active' : undefined}
-              onClick={() => togglePanel(ViewMenuPanel.LAYERS)}>
-              <StackSimple />
-            </button>
-
-            <DocumentNotesMenuIcon
-              active={panel === ViewMenuPanel.DOCUMENT_NOTES}
-              onSelect={() => togglePanel(ViewMenuPanel.DOCUMENT_NOTES)} />
-          </section>
-
-          {me && (
+          <div 
+            className="anno-menubar anno-desktop-overlay view-menu">
             <section>
-              <Avatar 
-                id={me.id}
-                name={me.appearance.label}
-                color={me.appearance.color} 
-                avatar={me.appearance.avatar} />
-            </section>
-          )}
-
-          {headerTransition((style, panel) => panel && (
-            <animated.section className="close" style={style}>
-              <button onClick={() => setPanel(undefined)}>
-                <X />
+              <button 
+                className={panel === ViewMenuPanel.ANNOTATIONS ? 'active' : undefined}
+                onClick={() => togglePanel(ViewMenuPanel.ANNOTATIONS)}>
+                <Chats />
               </button>
-            </animated.section>
-          ))}
-        </div>
 
-        {panelTransition((style, panel) => panel && (
-          <animated.aside style={style}>
-            {panel === ViewMenuPanel.ANNOTATIONS ? (
-              <AnnotationList 
-                i18n={props.i18n}
-                present={props.present} 
-                me={me}
-                policies={props.policies}
-                sorting={props.sorting}
-                tagVocabulary={props.tagVocabulary}
-                beforeSelect={props.beforeSelectAnnotation} />
-            ) : panel === ViewMenuPanel.LAYERS ? (
-              <LayersPanel
-                i18n={props.i18n}
-                layers={props.layers}
-                present={props.present}
-                onChangeStyle={props.onChangeAnnotationStyle} 
-                onChangeFilter={props.onChangeAnnotationFilter} />
-            ) : panel === ViewMenuPanel.DOCUMENT_NOTES ? props.defaultLayer && (
-              <DocumentNotesList 
-                i18n={props.i18n}
-                present={props.present}
-                policies={props.policies} 
-                tagVocabulary={props.tagVocabulary} />
-            ) : undefined}
-          </animated.aside>
-        ))}
-        
-      </DocumentNotes>
+              <button
+                className={panel === ViewMenuPanel.LAYERS ? 'active' : undefined}
+                onClick={() => togglePanel(ViewMenuPanel.LAYERS)}>
+                <StackSimple />
+              </button>
+
+              <DocumentNotesMenuIcon
+                active={panel === ViewMenuPanel.DOCUMENT_NOTES}
+                onSelect={() => togglePanel(ViewMenuPanel.DOCUMENT_NOTES)} />
+            </section>
+
+            {me && (
+              <section>
+                <Avatar 
+                  id={me.id}
+                  name={me.appearance.label}
+                  color={me.appearance.color} 
+                  avatar={me.appearance.avatar} />
+              </section>
+            )}
+
+            {headerTransition((style, panel) => panel && (
+              <animated.section className="close" style={style}>
+                <button onClick={() => setPanel(undefined)}>
+                  <X />
+                </button>
+              </animated.section>
+            ))}
+          </div>
+
+          {panelTransition((style, panel) => panel && (
+            <animated.aside style={style}>
+              {panel === ViewMenuPanel.ANNOTATIONS ? (
+                <AnnotationList 
+                  i18n={props.i18n}
+                  present={props.present} 
+                  me={me}
+                  policies={props.policies}
+                  sorting={props.sorting}
+                  tagVocabulary={props.tagVocabulary}
+                  beforeSelect={props.beforeSelectAnnotation} />
+              ) : panel === ViewMenuPanel.LAYERS ? (
+                <LayersPanel
+                  i18n={props.i18n}
+                  layers={props.layers}
+                  present={props.present}
+                  onChangeStyle={props.onChangeAnnotationStyle} 
+                  onChangeFilter={props.onChangeAnnotationFilter} />
+              ) : panel === ViewMenuPanel.DOCUMENT_NOTES ? props.defaultLayer && (
+                <DocumentNotesList 
+                  i18n={props.i18n}
+                  present={props.present}
+                  policies={props.policies} 
+                  tagVocabulary={props.tagVocabulary} />
+              ) : undefined}
+            </animated.aside>
+          ))}
+          
+        </DocumentNotes>
+      </FilterState>
     </div>
   )
